requirements.txt
Flask-Caching
langchain
langchain-groq 
langchain-community 
langchain-core 
langchain-cerebras
langchain-experimental
langgraph
flask
waitress
pandas
numpy
Flask-Cors
pydantic
pyyaml